combine workers
