+++ packages/browser-extension/src/modules/worker/ado/auth.ts
