+++ packages/browser-extension/src/modules/worker/handlers/handle-build-index.ts
