+++ packages/browser-extension/src/modules/worker/handlers/handle-import-index.ts
