+++ packages/browser-extension/src/modules/worker/handlers/handle-reset.ts
