+++ packages/browser-extension/src/modules/worker/handlers/handle-search.ts
