+++ packages/browser-extension/src/modules/worker/handlers/handle-sync.ts
