+++ packages/browser-extension/src/modules/worker/handlers/handle-test-connection.ts
