+++ packages/browser-extension/src/modules/worker/utils/diff.ts
