+++ packages/browser-extension/src/modules/worker/utils/iteration.ts
