+++ packages/browser-extension/src/modules/worker/utils/page.ts
