+++ packages/browser-extension/src/modules/worker/utils/search-index.ts
